# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: API for managing bank cards
  version: "1.0"
servers:
- url: http://127.0.0.1:8080
  description: Generated server url
tags:
- name: Authentication
  description: Authentication
- name: Cards
  description: Operations with cards
paths:
  /login:
    post:
      tags:
      - Authentication
      summary: Login
      description: Authentication to recieve a token
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/cards:
    get:
      tags:
      - Cards
      summary: Get all cards by filter
      description: Returns all cards information by filter and pageable. Available
        for admin and user roles only.
      operationId: getAllByFilter
      parameters:
      - name: expirationDateFrom
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: expirationDateTo
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
      - name: minBalance
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: maxBalance
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: userId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: createdFrom
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: createdTo
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: updatedFrom
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: updatedTo
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedModelCardReadDTO"
      security:
      - Bearer Authentication: []
    post:
      tags:
      - Cards
      summary: Create card
      description: Creates a new card. Available for admin role only.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardReadDTO"
      security:
      - Bearer Authentication: []
  /api/v1/cards/transfer:
    post:
      tags:
      - Cards
      summary: Transfer
      description: Transfer money between own cards. Available for user role only.
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardTransfer"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardReadDTO"
                  maxItems: 2
                  minItems: 2
                maxItems: 2
                minItems: 2
      security:
      - Bearer Authentication: []
  /api/v1/cards/{id}/status:
    patch:
      tags:
      - Cards
      summary: Update card status
      description: Updates card status. Available for admin role only.
      operationId: updateStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUpdateStatusDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardReadDTO"
      security:
      - Bearer Authentication: []
  /api/v1/cards/{id}/block:
    patch:
      tags:
      - Cards
      summary: Block card
      description: Blocks own card. Available for user role only.
      operationId: blockRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardReadDTO"
              example:
                id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                number: '**** **** **** 2820'
                expirationDate: 2025-12-23
                status: BLOCKED
                balance: 0
                userId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                createdAt: 2025-12-23T01:59:56.379Z
                updatedAt: 2025-12-23T01:59:56.379Z
      security:
      - Bearer Authentication: []
  /api/v1/cards/{id}:
    get:
      tags:
      - Cards
      summary: Get card by id
      description: Returns card information by its identifier. Available for admin
        and user roles only.
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardReadDTO"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
      security:
      - Bearer Authentication: []
    delete:
      tags:
      - Cards
      summary: Delete card
      description: Deletes card. Available for admin role only.
      operationId: deleteById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
      security:
      - Bearer Authentication: []
  /api/v1/cards/statuses:
    get:
      tags:
      - Cards
      summary: Get all statuses
      description: Returns all statuses. Available for admin role only.
      operationId: getAllStatuses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  statuses:
                    type: array
                    items:
                      type: string
                      enum:
                      - ACTIVE
                      - BLOCKED
                      - EXPIRED
              example:
                statuses:
                - ACTIVE
                - BLOCKED
                - EXPIRED
      security:
      - Bearer Authentication: []
components:
  schemas:
    AuthRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - password
      - username
    AuthResponse:
      type: object
      properties:
        token:
          type: string
    CardCreateDTO:
      type: object
      properties:
        expirationDate:
          type: string
          format: date
        userId:
          type: string
          format: uuid
      required:
      - expirationDate
      - userId
    CardReadDTO:
      type: object
      description: Data transfer object represent card
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
          description: Card number
          example: '**** **** **** 2820'
        expirationDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        balance:
          type: integer
          format: int64
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CardTransfer:
      type: object
      properties:
        fromCard:
          type: string
          format: uuid
        toCard:
          type: string
          format: uuid
        amount:
          type: integer
          format: int64
      required:
      - amount
      - fromCard
      - toCard
    CardUpdateStatusDTO:
      type: object
      properties:
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
      required:
      - status
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModelCardReadDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardReadDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
    DefaultHttpStatusCode:
      allOf:
      - $ref: "#/components/schemas/HttpStatusCode"
    ErrorInfo:
      type: object
      properties:
        httpStatus:
          oneOf:
          - $ref: "#/components/schemas/DefaultHttpStatusCode"
          - $ref: "#/components/schemas/HttpStatus"
        message:
          type: string
        printStackTrace:
          type: string
    HttpStatus:
      allOf:
      - $ref: "#/components/schemas/HttpStatusCode"
      enum:
      - 100 CONTINUE
      - 101 SWITCHING_PROTOCOLS
      - 102 PROCESSING
      - 103 EARLY_HINTS
      - 103 CHECKPOINT
      - 200 OK
      - 201 CREATED
      - 202 ACCEPTED
      - 203 NON_AUTHORITATIVE_INFORMATION
      - 204 NO_CONTENT
      - 205 RESET_CONTENT
      - 206 PARTIAL_CONTENT
      - 207 MULTI_STATUS
      - 208 ALREADY_REPORTED
      - 226 IM_USED
      - 300 MULTIPLE_CHOICES
      - 301 MOVED_PERMANENTLY
      - 302 FOUND
      - 302 MOVED_TEMPORARILY
      - 303 SEE_OTHER
      - 304 NOT_MODIFIED
      - 305 USE_PROXY
      - 307 TEMPORARY_REDIRECT
      - 308 PERMANENT_REDIRECT
      - 400 BAD_REQUEST
      - 401 UNAUTHORIZED
      - 402 PAYMENT_REQUIRED
      - 403 FORBIDDEN
      - 404 NOT_FOUND
      - 405 METHOD_NOT_ALLOWED
      - 406 NOT_ACCEPTABLE
      - 407 PROXY_AUTHENTICATION_REQUIRED
      - 408 REQUEST_TIMEOUT
      - 409 CONFLICT
      - 410 GONE
      - 411 LENGTH_REQUIRED
      - 412 PRECONDITION_FAILED
      - 413 PAYLOAD_TOO_LARGE
      - 413 REQUEST_ENTITY_TOO_LARGE
      - 414 URI_TOO_LONG
      - 414 REQUEST_URI_TOO_LONG
      - 415 UNSUPPORTED_MEDIA_TYPE
      - 416 REQUESTED_RANGE_NOT_SATISFIABLE
      - 417 EXPECTATION_FAILED
      - 418 I_AM_A_TEAPOT
      - 419 INSUFFICIENT_SPACE_ON_RESOURCE
      - 420 METHOD_FAILURE
      - 421 DESTINATION_LOCKED
      - 422 UNPROCESSABLE_ENTITY
      - 423 LOCKED
      - 424 FAILED_DEPENDENCY
      - 425 TOO_EARLY
      - 426 UPGRADE_REQUIRED
      - 428 PRECONDITION_REQUIRED
      - 429 TOO_MANY_REQUESTS
      - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
      - 451 UNAVAILABLE_FOR_LEGAL_REASONS
      - 500 INTERNAL_SERVER_ERROR
      - 501 NOT_IMPLEMENTED
      - 502 BAD_GATEWAY
      - 503 SERVICE_UNAVAILABLE
      - 504 GATEWAY_TIMEOUT
      - 505 HTTP_VERSION_NOT_SUPPORTED
      - 506 VARIANT_ALSO_NEGOTIATES
      - 507 INSUFFICIENT_STORAGE
      - 508 LOOP_DETECTED
      - 509 BANDWIDTH_LIMIT_EXCEEDED
      - 510 NOT_EXTENDED
      - 511 NETWORK_AUTHENTICATION_REQUIRED
    HttpStatusCode:
      type: object
      properties:
        error:
          type: boolean
        is4xxClientError:
          type: boolean
        is5xxServerError:
          type: boolean
        is1xxInformational:
          type: boolean
        is2xxSuccessful:
          type: boolean
        is3xxRedirection:
          type: boolean
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT